pyproject.toml
[project]
name = "xrpl-quality-agent"
version = "0.1.0"
description = "Scenario runner + telemetry + summaries (portfolio)"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rich>=13.7",
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.6",
]

[project.scripts]
qe = "qe.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","I","UP","B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["qe*"]
exclude = ["tests*", "scenarios*"]

